server source

+++ src/pages/Persons/Attendance/index.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react';
 import {Animate} from '@/contexts';
-import {Header, Icon, Search } from "@/components";
+import {Header, Icon, Search, Checkbox } from "@/components";
 import classNames from 'classnames';
 import {useSelector} from 'react-redux';
 import cls from './style.module.scss';
@@ -12,11 +12,10 @@ import {unixData} from './h_date.js';
 
 function Attendance (props) {
 	const attendance = useSelector(state => state.persons.attendance);
+	const {isAdmin} = useSelector(state => state.admin);
 	const [searchValue, setSearchValue] = useState('');
 	const [filteredAttendance, setFilteredAttendance] = useState([]);
 	const actions = useActions();
-
-
 
 	useEffect(()=>{
 		let val = searchValue.trim().toLowerCase();
@@ -32,6 +31,10 @@ function Attendance (props) {
 
 	const getDate = (date) => unixData(date).made.ru.contextDateMonthYear;
 
+	const changeCheckbox = (e, id) => {
+		let value = e.target.checked;
+		actions.changeAttendanceGraph({graph: value, id});
+	} 
 
 
 	const renderItem = el => (
@@ -45,10 +48,14 @@ function Attendance (props) {
 					exit={{ opacity: 0, x: -100 }}
 				>
 					<h4>{el.name}</h4>
-					<p>{getDate(el.date)}</p>
-					<div className={cls.item__btns}>
-						<button onClick={() => deleteItem(el)} className={cls.item__btn}><Icon name='trash'/></button>
+ 					<div className={cls.item__btns}>
+						<button style={{display: isAdmin ? "flex" : "none"}} onClick={() => deleteItem(el)} className={cls.item__btn}><Icon name='trash'/></button>
 						<Link to={'/persons/info-attendance/' + el.id} className={cls.item__btn}><Icon name='scale'/></Link>
+					</div>
+					
+					<div className={cls.item__info}>
+						<label style={{pointerEvents: isAdmin ? 'auto' : 'none'}}> <span>Мусхаф</span> <Checkbox size={10} type='slider' onChange={e => changeCheckbox(e, el.id)} defaultChecked={el.graph}/></label>
+						<p>{getDate(el.date)}</p>
 					</div>
 				</motion.li>
 			}
@@ -63,7 +70,7 @@ function Attendance (props) {
 					<Header title="Посещаемость" nav='/persons'/>
 					<div className={cls.head}>
 						<Search setter={setSearchValue} />
-						<Link to='/persons/add-attendance' className={cls.add_btn}> <Icon name='books' /> </Link>
+						<Link style={{display: isAdmin ? "flex" : "none"}} to='/persons/add-attendance' className={cls.add_btn}> <Icon name='books' /> </Link>
 					</div>
 				</div>
